xdp: tx_loop: allow overriding MACs and IP (#7285)
* xdp: use AsRef for addresses too

This allows passing Arc<Vec<_>> when sending to a static list of peers.

* xdp: tx_loop: allow overriding MACs and IP

This is useful for sending via bonded NICs, skipping the bond device and
targeting the slaved NICs directly so they can be used in zero copy
mode.

Slaved NICs don't have an IP assigned to them, so the source IP can't be
discovered but must be passed explicitly (eg the bond IP).

Overriding MACs can be useful while testing to ensure frames don't leave
the switch/metro.