wip: refactor: do not store epoch related state in `CertificatePool` (#360)
Instead of caching epoch related state in the CertificatePool, pass it into the pool. This removes dependencies between banks and other related structs like the RootBankCache, and BankForks, etc.

Also updates cleanup() to use the highest finalized slot instead of the root slot which is in line with the paper.