introduce cancel to streamer (#8025)
When the exit signal is set to true, the streamer's tasks are not exit because not all of them track exit.
In this PR, we introduce CancellationToken instead of exit for nonblocking::streamer. This is the only way to gracefully close connection tasks.
It introduces spawn_server_with_cancel and deprecates spawn_server which uses exit. So all the tests has been changed to reflect this change.

(cherry picked from commit a0e531eb04980ff5aba9e662159fa2ca0874a6b4)