fast leader handover: error on bad `UpdateParent` during shred ingest (#605)
When disseminating an `UpdateParent` component, we need to ensure that
the parent change passes validity checks with respect to the
`BlockHeader`.

We explicitly cover the four cases:
- `UpdateParent` observed before `BlockHeader` and...
- ... `UpdateParent` and `BlockHeader` are in the same shred insertion
batch
- ... `UpdateParent` and `BlockHeader` are in different shred insertion
batches

- `BlockHeader` observed before `UpdateParent` and...
- ... `UpdateParent` and `BlockHeader` are in the same shred insertion
batch
- ... `UpdateParent` and `BlockHeader` are in different shred insertion
batches

These checks are efficient, since we can detect the presence of a
`BlockHeader` in a shred's payload without having to touch blockstore.