bcl: pack the GenesisCertificate into the first alpenglow block(s) (#600)
The direct child(ren) of the alpenglow genesis block must contain the
Alpenglow Genesis certificate.
This is critical in order to update the off curve account for use in
restarting post migration.

Send the genesis certificate block marker for these blocks. Next PR will
plug in the processing of it during replay.