migration: hold shred lock while purging blockstore (#610)
Noticed some very rare flakiness in migration local-cluster tests where
one node was running behind:
- The slow node attempts to enable alpenglow and purges its blockstore
columns
- During the process of purging a new shred from an Alpenglow block is
ingested
- Since the purge is not being done atomically, the `SlotMeta` can end
up in a state that is inconsistent
- Specifically seeing situations where `parent_slot` is set for the
child block, but `next_slots` for the parent does not point to the child
block.

The original purge logic was copied from the startup process that cleans
up shreds with incorrect shred versions. This is sound because window
service has not yet started so there's no race.

Instead replace it with logic similar to dump & repair, which takes care
to acquire the shred insertion lock to ensure that window service is not
inserting shreds at the same time.