blockstore: compute and populate DoubleMerkleRootMeta (#613)
Adds an unoptimized double merkle root computation to get the ball
rolling.

Computes and stores the double merkle root and associated proofs once
the block is full.
This places all the blockstore lookups in the hot path (computation will
run once block has been frozen).

Future optimizations:
1. Join nodes as each FEC set comes in, no need to wait until the end
and perform up to 1000 blockstore lookups for FEC set roots, this can be
done in line when populating new MerkleRootMetas
2. Generate proofs only when the first repair request comes in. The
majority of blocks should not be repaired - we can defer this out of the
hot path.

Next PR will actually call this fn and set block id in either replay or
broadcast