Call handle_event directly in tests. (#631)
Call handle_event directly in tests.

The benefit is that we don't rely on timing any more.

But we need to promise never adding logic outside handle_event.

This is a downstream of https://github.com/anza-xyz/agave/pull/9010. But
cherry-picking is too painful because the two diverged, so I just
applied all changes.