fast leader handover: `UpdateParent` events in `BlockComponentProcessor` (#628)
We need to slightly rework how we process `BlockComponent`s in
`BlockComponentProcessor` to work with fast leader handover:

**Case (1): the first component we process is a `BlockHeader`**
- Process one component at a time
- If we run into an `UpdateParent`, then return an `AbandonedBank` error
- In a separate PR, replay will clear the bank, resetting
`BlockComponentProcessor`'s state
- `generate_new_bank_forks` picks up the new bank on the next iteration
of the replay loop
- We intentionally don't update the progress bar, resuming from the
`UpdateParent`. We then end up in case (2) below.

**Case (2): the first component we process is an `UpdateParent`**
- Because the *first* component we process is an `UpdateParent`, we
don't return an `AbandonedBank` error. Instead, we keep processing
events.
- We eventually get to the block footer.
- We run `is_final` checks at the end, and we're all set.

This PR implements the bullet points for the above logic.